--- libs/wrappers/pyproject.toml
@@ -1,13 +1,0 @@
-[tool.poetry]
-name = "wrappers"
-version = "0.1.0"
-description = "Python wrappers for various linux cli tools."
-authors = ["Felix Trautwein <mail@felixtrautwein.de>"]
-
-[tool.poetry.dependencies]
-python="^3.10"
-paramiko="^3.3.1"
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.core.masonry.api"
