+++ tools/devtools/src/devtools/wrappers/docker.py
@@ -2,7 +2,7 @@
 
 from pathlib import PurePath
 
-from wrappers.linux_wrapper import linux
+from devtools.wrappers import linux
 
 
 def install_docker() -> list[str]:
